requirements.txt
FLask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
git+https://github.com/kota113/Bard-API.git@refresh-cookies-automatically
genanki==0.13.0
gunicorn==21.2.0
google-generativeai